Simplify test hierarchy
