+++ src/main.js
@@ -2,19 +2,19 @@
 
 const path = require('path')
 const Bluebird = require('bluebird')
+const os = require('os')
+const git = require('nodegit')
+const fse = require('fs-extra')
+const chai = require('chai')
+const chaiAsPromised = require('chai-as-promised')
+chai.use(chaiAsPromised)
+
 const resinio = require('./components/resinio/sdk')
 const resinos = require('./components/resinos/simple')
 const writer = require('./components/writer/etcher')
-const os = require('os')
 
-global.options = require(path.join(__dirname, '../user.json'))
-global.provDevice = null
-
-/* TODO: Re-enable
-const importSuite = (name, testPath) => {
-  describe(name, require(testPath)().describe)
-}
-*/
+const options = require(path.join(__dirname, '../user.json'))
+let provDevice = null
 
 describe('Test ResinOS', function () {
   this.timeout(600000)
@@ -36,9 +36,9 @@ describe('Test ResinOS', function () {
 
       return Bluebird.resolve()
     }).then(() => {
-      return resinio.createApplication(applicationName, global.options.deviceType)
+      return resinio.createApplication(applicationName, options.deviceType)
     }).then(() => {
-      return resinio.downloadDeviceTypeOS(global.options.deviceType, global.options.version, this.imagePath)
+      return resinio.downloadDeviceTypeOS(options.deviceType, options.version, this.imagePath)
     }).then(() => {
       return resinio.getApplicationOSConfiguration(applicationName, configuration).then((applicationConfiguration) => {
         return resinos.injectResinConfiguration(this.imagePath, applicationConfiguration)
@@ -46,10 +46,61 @@ describe('Test ResinOS', function () {
     }).then(() => {
       return resinos.injectNetworkConfiguration(this.imagePath, configuration)
     }).then(() => {
-      return writer.writeImage(this.imagePath, global.options.disk)
+      return writer.writeImage(this.imagePath, options.disk)
     })
   })
 
-  // TODO: importSuite(`Device`, './resin/device.js')
-  // TODO: importSuite('Container', './resin/container.js')
+  // eslint-disable-next-line prefer-arrow-callback
+  xit('Device should become online', function () {
+    this.retries(50)
+
+    return Bluebird.delay(10000)
+      .return(process.env.APPLICATION_NAME)
+      .then(resinio.getApplicationDevices)
+      .then((devices) => {
+        chai.expect(devices).to.have.length(1)
+        chai.expect(devices).to.be.instanceof(Array)
+        return devices[0]
+      })
+      .then((device) => {
+        provDevice = device.id
+        return resinio.isDeviceOnline(device)
+      }).then((isOnline) => {
+        return chai.expect(isOnline).to.be.true
+      })
+  })
+
+  // eslint-disable-next-line prefer-arrow-callback
+  xit(`Device should report hostOS version: ${options.version}`, function () {
+    return resinio.getDeviceHostOSVersion(provDevice).then((version) => {
+      return chai.expect(version).to.equal('Resin OS 2.0.6+rev3')
+    })
+  })
+
+  // eslint-disable-next-line prefer-arrow-callback
+  xit('should push an application', function () {
+    let repository = null
+
+    const repoPath = path.join(os.tmpdir(), 'test')
+    return fse.remove(repoPath).then(() => {
+      return git.Clone(options.gitAppURL, repoPath)
+    }).then((repo) => {
+      repository = repo
+      return resinio.getApplicationGitRemote(process.env.APPLICATION_NAME)
+    }).then((remote) => {
+      return git.Remote.create(repository, 'resin', remote)
+    }).then((remote) => {
+      return remote.push([ 'refs/heads/master:refs/heads/master' ], {
+        callbacks: {
+          credentials: (url, user) => {
+            return git.Cred.sshKeyFromAgent(user)
+          }
+        }
+      })
+    }).then(() => {
+      return resinio.getDeviceCommit(provDevice).then((commit) => {
+        return chai.expect(commit).to.have.length(40)
+      })
+    })
+  })
 })
